requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
graphene-django
drf-yasg
psycopg2-binary
gunicorn
whitenoise
